+++ ql/experimental/barrieroption/all.hpp
@@ -3,6 +3,7 @@
 
 #include <ql/experimental/barrieroption/analyticdoublebarrierbinaryengine.hpp>
 #include <ql/experimental/barrieroption/analyticdoublebarrierengine.hpp>
+#include <ql/experimental/barrieroption/mcdoublebarrierengine.hpp>
 #include <ql/experimental/barrieroption/binomialdoublebarrierengine.hpp>
 #include <ql/experimental/barrieroption/discretizeddoublebarrieroption.hpp>
 #include <ql/experimental/barrieroption/doublebarrieroption.hpp>
